Getting lightning running
